backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
openai==1.12.0
pinecone-client==3.2.2
pydantic==2.6.1
python-multipart==0.0.6